Refactor type system
