--- src/lib/types/addedfriends.ts
@@ -1,11 +1,0 @@
-export type AddedFriend = {
-    id: string;
-    name: string;
-    username: string;
-    pfp?: string;
-    nowPlaying?: {
-        title: string;
-        artist: string;
-        album: string;
-    };
-}
