--- src/lib/types/album.ts
@@ -1,9 +1,0 @@
-export type Album = {
-	id: string;
-	name: string;
-	artist: string;
-	image: any;
-	year: number;
-	genre?: string;
-	tracks?: String[];
-};
