--- src/lib/types/artist.ts
@@ -1,7 +1,0 @@
-export type Artist = {
-	id: string;
-	name: string;
-	image?: any;
-	tracks?: String[];
-	albums?: String[];
-};
