--- src/lib/types/playlist.ts
@@ -1,7 +1,0 @@
-export type Playlist = {
-	id: string;
-	name: string;
-	description: string;
-	tracks?: String[];
-	image?: any;
-};
