--- src/lib/types/preq.ts
@@ -1,4 +1,0 @@
-export type PendingRequest = {
-    user_id: string,
-    friend_id: string
-};
