--- src/lib/types/song.ts
@@ -1,14 +1,0 @@
-export type Song = {
-	id: string;
-	title: string;
-	artist: string;
-	album: string;
-	year: number;
-	genre?: string;
-	fileName: string;
-	duration: number;
-	image?: any;
-	trackNumber: number;
-	disk: number;
-	ext: string;
-};
