--- src/lib/types/user.ts
@@ -1,6 +1,0 @@
-export type User = {
-	id: string;
-	username: string;
-	name?: string;
-	pfp?: any;
-};
