+++ src/routes/account/+page.svelte
@@ -3,7 +3,7 @@
 	import Button from '$lib/components/ui/button/button.svelte';
 	import Input from '$lib/components/ui/input/input.svelte';
 	import { SavedUser, title } from '$lib/store';
-	import type { User } from '$lib/types/user';
+	import type { User } from '$lib/types';
 	import { User as UserIcon } from 'lucide-svelte';
 	import { onMount } from 'svelte';
